User service almost fixed full

+++ UserService/pkg/db/repository/utils.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 )
 
-func BanRecension(recensionId uint) error {
+func BanRecension(recensionId uint64) error {
 
 	endpoint := "http://localhost:50005/api/recensions/" + strconv.Itoa(int(recensionId))
 	req, _ := http.NewRequest(http.MethodDelete, endpoint, nil)
